Refactoring folders project
